♻️ Refactoring pull request https://github.com/it-at-m/mobidam-mobilithek-eai/pull/18
