+++ src/main/java/de/muenchen/mobidam/mobilithek/EreignisTyp.java
@@ -1,6 +1,6 @@
 /*
  * The MIT License
- * Copyright © 2023 Landeshauptstadt München | it@M
+ * Copyright © 2024 Landeshauptstadt München | it@M
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
