+++ src/main/java/de/muenchen/mobidam/mobilithek/InterfaceMessageFactory.java
@@ -1,3 +1,25 @@
+/*
+ * The MIT License
+ * Copyright © 2024 Landeshauptstadt München | it@M
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 package de.muenchen.mobidam.mobilithek;
 
 import de.muenchen.mobidam.Constants;
@@ -22,7 +44,6 @@ public class InterfaceMessageFactory {
         dto.setZeitstempel(LocalDateTime.now());
         dto.setProzessId(exchange.getExchangeId());
         dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_DATA).getMobidamSstId());
-        dto.setInfo("Interface start.");
         exchange.getMessage().setBody(dto);
     }
 
@@ -50,17 +71,6 @@ public class InterfaceMessageFactory {
         exchange.getMessage().setBody(dto);
     }
 
-    public void parkRideDataWarn(Exchange exchange) {
-
-        var dto = new DatentransferCreateDTO();
-        dto.setEreignis(EreignisTyp.WARNUNGEN.name());
-        dto.setZeitstempel(LocalDateTime.now());
-        dto.setProzessId(exchange.getExchangeId());
-        dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_DATA).getMobidamSstId());
-        dto.setInfo("Interface is not activated");
-        exchange.getMessage().setBody(dto);
-    }
-
     public void parkRideDataEnd(Exchange exchange) {
 
         var dto = new DatentransferCreateDTO();
@@ -68,7 +78,6 @@ public class InterfaceMessageFactory {
         dto.setZeitstempel(LocalDateTime.now());
         dto.setProzessId(exchange.getExchangeId());
         dto.setSchnittstelle(interfaceProps.getInterfaces().get(Constants.PARK_RIDE_DATA).getMobidamSstId());
-        dto.setInfo("Interface end");
         exchange.getMessage().setBody(dto);
     }
