--- src/main/resources/application-local.yml
@@ -1,12 +1,0 @@
-server:
-  port: 8083
-
-camel:
-  springboot:
-    #    Sollte sich der Camel Main Thread nach dem Start wieder beenden, folgende Property aktivieren (http://camel.apache.org/spring-boot.html).
-    main-run-controller: true
-    tracing: false
-
-info:
-  application:
-    version: @project.version@-local
